+++ framework/js/onsen.js
@@ -234,6 +234,7 @@ window.ons = (function(){
   waitDeviceReady();
   waitOnsenUILoad();
   initAngularModule();
+  changeHammerDefault();
 
   return ons;
 
@@ -282,6 +283,11 @@ window.ons = (function(){
 
       ons.$compile = $compile;
     });
+  }
+
+  //Change the default touchAction of Hammer.js, needed for Windows Phone app
+  function changeHammerDefault() {
+    Hammer.defaults.behavior.touchAction = 'none';
   }
 
   function initKeyboardEvents() {