+++ framework/views/navigator.js
@@ -329,8 +329,9 @@ limitations under the License.
 
       /**
        * @param {Object} options pushPage()'s options parameter
+       * @param {NavigatorTransitionAnimator} [defaultAnimator]
        */
-      _getAnimatorOption: function(options) {
+      _getAnimatorOption: function(options, defaultAnimator) {
         var animator = null;
 
         if (options.animation instanceof NavigatorTransitionAnimator) {
@@ -346,7 +347,7 @@ limitations under the License.
         }
 
         if (!animator) {
-          animator = NavigatorView._transitionAnimatorDict['default'];
+          animator = defaultAnimator || NavigatorView._transitionAnimatorDict['default'];
         }
 
         if (!(animator instanceof NavigatorTransitionAnimator)) {
@@ -487,7 +488,7 @@ limitations under the License.
        */
       popPage: function(options) {
         options = options || {};
-        
+
         this._doorLock.waitUnlock(function() {
           if (this.pages.length <= 1) {
             throw new Error('NavigatorView\'s page stack is empty.');
@@ -532,7 +533,30 @@ limitations under the License.
         }.bind(this);
 
         this._isPopping = true;
-        leavePage.options.animator.pop(enterPage, leavePage, callback);
+
+        var animator = this._getAnimatorOption(options, leavePage.options.animator);
+        animator.pop(enterPage, leavePage, callback);
+      },
+
+      /**
+       * Replaces the current page with the specified one.
+       *
+       * @param {String} page
+       * @param {Object} [options]
+       */
+      replacePage: function(page, options) {
+        options = options || {};
+
+        var onTransitionEnd = options.onTransitionEnd || function() {};
+
+        options.onTransitionEnd = function() {
+          if (this.pages.length > 1) {
+            this.pages[this.pages.length - 2].destroy();
+          }
+          onTransitionEnd();
+        }.bind(this);
+
+        this.pushPage(page, options);
       },
 
       /**
@@ -569,7 +593,7 @@ limitations under the License.
        * Use this method to access options passed by pushPage() or resetToPage() method.
        * eg. ons.navigator.getCurrentPage().options
        *
-       * @return {Object} 
+       * @return {Object}
        */
       getCurrentPage: function() {
         return this.pages[this.pages.length - 1];