+++ framework/views/tabbarView.js
@@ -158,12 +158,12 @@ limitations under the License.
         var previousTabItem = this._tabItems[this.getActiveTabIndex()];
         var selectedTabItem = this._tabItems[index];
 
-        if(typeof selectedTabItem.noReload !== 'undefined' && 
-            index == this.getActiveTabIndex()) {
+        if((typeof selectedTabItem.noReload !== 'undefined' || typeof selectedTabItem.isPersistent()) &&
+            index === this.getActiveTabIndex()) {
           this.emit('reactive', {
             index: index,
             tabItem: selectedTabItem,
-          }); 
+          });
           return false;
         }